+++ src/host/main.cpp
@@ -248,23 +248,23 @@ int main(int argc, char** argv)
       "election.")
     ->capture_default_str();
 
-  size_t pbft_view_change_timeout = 5000;
+  size_t bft_view_change_timeout = 5000;
   app
     .add_option(
-      "--pbft_view-change-timeout-ms",
-      pbft_view_change_timeout,
-      "Pbft view change timeout in milliseconds. If a backup does not receive "
+      "--bft_view-change-timeout-ms",
+      bft_view_change_timeout,
+      "bft view change timeout in milliseconds. If a backup does not receive "
       "the pre-prepare message for a request forwarded to the primary after "
       "this "
       "timeout, the backup triggers a new view change.")
     ->capture_default_str();
 
-  size_t pbft_status_interval = 100;
+  size_t bft_status_interval = 100;
   app
     .add_option(
-      "--pbft-status-interval-ms",
-      pbft_status_interval,
-      "Pbft status timer interval in milliseconds. All pbft nodes send "
+      "--bft-status-interval-ms",
+      bft_status_interval,
+      "bft status timer interval in milliseconds. All bft nodes send "
       "messages "
       "containing their status to all other known nodes at regular intervals "
       "defined by this timer interval.")
@@ -362,8 +362,8 @@ int main(int argc, char** argv)
     *start,
     members_info,
     "--member-info",
-    "Initial consortium members information (public identity,encryption public "
-    "key,member data)")
+    "Initial consortium members information "
+    "(member_cert.pem[,member_enc_pubk.pem[,member_data.json]])")
     ->required();
 
   std::optional<size_t> recovery_threshold = std::nullopt;
@@ -372,7 +372,7 @@ int main(int argc, char** argv)
       "--recovery-threshold",
       recovery_threshold,
       "Number of member shares required for recovery. Defaults to total number "
-      "of initial consortium members.")
+      "of initial consortium members with a public encryption key.")
     ->check(CLI::PositiveNumber)
     ->type_name("UINT");
 
@@ -453,20 +453,36 @@ int main(int argc, char** argv)
 
     if (*start)
     {
+      // Count members with public encryption key as only these members will be
+      // handed a recovery share.
+      // Note that it is acceptable to start a network without any member having
+      // a recovery share. The service will check that at least one recovery
+      // member is added before the service can be opened.
+      size_t members_with_pubk_count = 0;
+      for (auto const& mi : members_info)
+      {
+        if (mi.enc_pubk_file.has_value())
+        {
+          members_with_pubk_count++;
+        }
+      }
+
       if (!recovery_threshold.has_value())
       {
         LOG_INFO_FMT(
-          "--recovery-threshold unset. Defaulting to number of initial "
-          "consortium members ({}).",
-          members_info.size());
-        recovery_threshold = members_info.size();
+          "Recovery threshold unset. Defaulting to number of initial "
+          "consortium members with a public encryption key ({}).",
+          members_with_pubk_count);
+        recovery_threshold = members_with_pubk_count;
       }
-      else if (recovery_threshold.value() > members_info.size())
+      else if (recovery_threshold.value() > members_with_pubk_count)
       {
         throw std::logic_error(fmt::format(
-          "--recovery-threshold cannot be greater than total number "
-          "of initial consortium members (specified via --member-info "
-          "options)"));
+          "Recovery threshold ({}) cannot be greater than total number ({})"
+          "of initial consortium members with a public encryption "
+          "key (specified via --member-info options)",
+          recovery_threshold.value(),
+          members_with_pubk_count));
       }
     }
 
@@ -623,8 +639,8 @@ int main(int argc, char** argv)
     CCFConfig ccf_config;
     ccf_config.consensus_config = {raft_timeout,
                                    raft_election_timeout,
-                                   pbft_view_change_timeout,
-                                   pbft_status_interval};
+                                   bft_view_change_timeout,
+                                   bft_status_interval};
     ccf_config.signature_intervals = {sig_tx_interval, sig_ms_interval};
     ccf_config.node_info_network = {rpc_address.hostname,
                                     public_rpc_address.hostname,
@@ -643,6 +659,14 @@ int main(int argc, char** argv)
 
       for (auto const& m_info : members_info)
       {
+        std::optional<std::vector<uint8_t>> public_encryption_key_file =
+          std::nullopt;
+        if (m_info.enc_pubk_file.has_value())
+        {
+          public_encryption_key_file =
+            files::slurp(m_info.enc_pubk_file.value());
+        }
+
         nlohmann::json md = nullptr;
         if (m_info.member_data_file.has_value())
         {
@@ -651,9 +675,7 @@ int main(int argc, char** argv)
         }
 
         ccf_config.genesis.members_info.emplace_back(
-          files::slurp(m_info.cert_file),
-          files::slurp(m_info.enc_pub_file),
-          md);
+          files::slurp(m_info.cert_file), public_encryption_key_file, md);
       }
       ccf_config.genesis.gov_script = files::slurp_string(gov_script);
       ccf_config.genesis.recovery_threshold = recovery_threshold.value();