+++ src/node/entities.h
@@ -64,8 +64,10 @@ namespace ccf
     static constexpr auto MEMBERS = "public:ccf.gov.members";
     static constexpr auto MEMBER_ACKS = "public:ccf.gov.member_acks";
     static constexpr auto MEMBER_CERT_DERS = "public:ccf.gov.member_cert_ders";
+    static constexpr auto MEMBER_DIGESTS = "public:ccf.gov.member_digests";
     static constexpr auto USERS = "public:ccf.gov.users";
     static constexpr auto USER_CERT_DERS = "public:ccf.gov.user_cert_ders";
+    static constexpr auto USER_DIGESTS = "public:ccf.gov.user_digests";
     static constexpr auto NODES = "public:ccf.gov.nodes";
     static constexpr auto VALUES = "public:ccf.gov.values";
     static constexpr auto CONSENSUS = "public:ccf.gov.consensus";