+++ tests/infra/runner.py
@@ -19,10 +19,10 @@ logging.getLogger("paramiko").setLevel(logging.WARNING)
 
 def minimum_number_of_local_nodes(args):
     """
-    If we are using pbft then we need to have 4 nodes. CFT will run with 1 nodes, unless it expects a backup
+    If we are using bft then we need to have 3 nodes. CFT will run with 1 nodes, unless it expects a backup
     """
     if args.consensus == "bft":
-        return 4
+        return 3
 
     if args.send_tx_to == "backups":
         return 2