+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.15.0]
+
+### Added
+
+- Support for non-recovery members: only members with an associated public encryption key are handed recovery shares (#1866).
+- AFT consensus verify entry validity (#1864).
+- JWT validation in forum sample app (#1867).
+- JavaScript endpoints OpenAPI definition is now included in `/api` (#1874).
+
+### Changed
+
+- The `keyId` field in the Authorization header must now be set to the hex-encoded SHA-256 digest of the corresponding member certificate encoded in PEM format. The `scurl.sh` script and Python client have been modified accordingly. `scurl.sh` can be run with `DISABLE_CLIENT_AUTH=1` (equivalent `disable_client_auth=False` argument to Python client) to issue signed requests without session-level client authentication (#1870).
+- Governance endpoints no longer require session-level client authentication matching a member identity, the request signature now serves as authentication. The purpose of this change is to facilitate member key storage in systems such as HSMs (#1870).
+- Support for [hs2019 scheme](https://tools.ietf.org/html/draft-cavage-http-signatures-12) for HTTP signatures (#1872).
+  - `ecdsa-sha256` scheme will be deprecated in the next release.
+
 ## [0.14.3]
 
 ### Added
@@ -507,6 +523,7 @@ Some discrepancies with the TR remain, and are being tracked under https://githu
 
 Initial pre-release
 
+[0.15.0]: https://github.com/microsoft/CCF/releases/tag/ccf-0.15.0
 [0.14.3]: https://github.com/microsoft/CCF/releases/tag/ccf-0.14.3
 [0.14.2]: https://github.com/microsoft/CCF/releases/tag/ccf-0.14.2
 [0.14.1]: https://github.com/microsoft/CCF/releases/tag/ccf-0.14.1